test: :white_check_mark: Adiciona novos testes automatizados
